YAML format

+++ app/commands/repository/filesystem/destroy.rb
@@ -5,8 +5,10 @@ module Repository
     ##
     # Destroy repository directory
     #
-    class Destroy < RepoCommand
+    class Destroy < YAMLCommand
       def execute
+        raise OpenWebslides::RepoDoesNotExistError unless Dir.exist? repo_path
+
         FileUtils.rm_r repo_path, :secure => true
       end
     end
